Created a lot of functions for d conversion for negative numbers

+++ mini_libft/ft_abs.c
@@ -6,7 +6,7 @@
 /*   By: mbutt <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/09 19:58:09 by mbutt             #+#    #+#             */
-/*   Updated: 2019/08/09 20:00:01 by mbutt            ###   ########.fr       */
+/*   Updated: 2019/08/28 20:09:59 by mbutt            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 ** Return Value: Returns a postive value.
 */
 
-int		ft_abs(int num)
+int_fast64_t	ft_abs(int_fast64_t num)
 {
 	if (num < 0)
 		num = -1 * num;
